+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Added completions for:
  -
 - Improved completions for `killall` (#4052), `ln` (#4090) and `zypper` (#4079).
+- Implemented `string lower` and `string upper` (#4080).
 
 ---